# onboarding/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Interactive onboarding platform: tours, immutable versions, resolve and analytics.",
        "title": "Cluer Onboarding",
        "contact": {},
        "version": "0.1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/tours": {
            "get": {
                "description": "Returns every tour of an app with the status of its current version",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tours"
                ],
                "summary": "List tours",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "App ID",
                        "name": "appId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.Tour"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a tour together with its first draft version",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tours"
                ],
                "summary": "Create a new tour",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "App ID",
                        "name": "appId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Tour object",
                        "name": "tour",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Tour"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Tour"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errs.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/tours/{tourId}/hints": {
            "get": {
                "description": "Returns the hints of the tour's draft version, ordered by step",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hints"
                ],
                "summary": "List draft hints",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Tour ID",
                        "name": "tourId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.Hint"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/errs.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "Appends a hint to the tour's draft version",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hints"
                ],
                "summary": "Create a hint",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Tour ID",
                        "name": "tourId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Hint object",
                        "name": "hint",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Hint"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Hint"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.HTTPError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/errs.HTTPError"
                        }
                    }
                }
            }
        },
        "/tours/{tourId}/hints/order": {
            "put": {
                "description": "Takes the full list of hint ids in their new order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hints"
                ],
                "summary": "Reorder hints",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Tour ID",
                        "name": "tourId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Ordered hint ids",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ReorderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.Hint"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.HTTPError"
                        }
                    }
                }
            }
        },
        "/tours/{tourId}/hints/{hintId}": {
            "delete": {
                "description": "Removes a hint from the draft and closes the gap in step numbering",
                "tags": [
                    "Hints"
                ],
                "summary": "Delete a hint",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Tour ID",
                        "name": "tourId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Hint ID",
                        "name": "hintId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errs.HTTPError"
                        }
                    }
                }
            },
            "patch": {
                "description": "Edits a hint of the draft version; hints of published versions are immutable",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hints"
                ],
                "summary": "Update a hint",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Tour ID",
                        "name": "tourId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Hint ID",
                        "name": "hintId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Hint object",
                        "name": "hint",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Hint"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Hint"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/errs.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/users/me": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Returns the identity behind the presented id token: subject, profile fields and IdP groups.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get the current administrator",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.GetCurrentAdminResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errs.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errs.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "errs.HTTPError": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Shorthand error",
                    "type": "string"
                },
                "message": {
                    "description": "Pretty error message",
                    "type": "string"
                }
            }
        },
        "request.Audience": {
            "type": "object",
            "properties": {
                "max_shows": {
                    "type": "integer",
                    "example": 3
                },
                "only_new": {
                    "type": "boolean",
                    "example": false
                },
                "rules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.AudienceRule"
                    }
                },
                "show_once": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "request.Hint": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "Здесь вы можете изменить параметры аккаунта"
                },
                "created_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "expected_input": {
                    "type": "string",
                    "example": "John Doe"
                },
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "input_placeholder": {
                    "type": "string",
                    "example": "Введите ваше имя"
                },
                "media_url": {
                    "type": "string",
                    "example": "https://cdn.example.com/onboarding/settings.mp4"
                },
                "page_path": {
                    "description": "Пусто — подсказка живёт на той же странице, где тур стартовал,\nто есть наследует target_path версии.",
                    "type": "string",
                    "example": "/dashboard"
                },
                "placement": {
                    "type": "string",
                    "enum": [
                        "top",
                        "bottom",
                        "center",
                        "left-top",
                        "right-top",
                        "left",
                        "right"
                    ],
                    "example": "bottom"
                },
                "required": {
                    "type": "boolean",
                    "example": false
                },
                "selector": {
                    "type": "string",
                    "example": "#settings-btn"
                },
                "spotlight": {
                    "type": "boolean",
                    "example": true
                },
                "step": {
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "example": "Нажмите кнопку настроек"
                },
                "tour_version_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "updated_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "wait_for_selector": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "request.ReorderRequest": {
            "type": "object",
            "required": [
                "hint_ids"
            ],
            "properties": {
                "hint_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "request.Tour": {
            "type": "object",
            "properties": {
                "app_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "audience": {
                    "$ref": "#/definitions/request.Audience"
                },
                "created_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "description": {
                    "type": "string",
                    "example": "Тур по основным функциям дашборда"
                },
                "enabled": {
                    "type": "boolean",
                    "example": true
                },
                "hints": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.Hint"
                    }
                },
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "priority": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "draft",
                        "published",
                        "archived"
                    ],
                    "example": "draft"
                },
                "target_path": {
                    "type": "string",
                    "example": "/dashboard"
                },
                "title": {
                    "type": "string",
                    "example": "Добро пожаловать в систему"
                },
                "trigger_config": {
                    "$ref": "#/definitions/request.TriggerConfig"
                },
                "trigger_type": {
                    "type": "string",
                    "enum": [
                        "on_load",
                        "delay",
                        "exit_intent",
                        "manual",
                        "scroll_depth",
                        "inactivity",
                        "element_visible"
                    ],
                    "example": "on_load"
                },
                "updated_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "version": {
                    "type": "integer",
                    "example": 7
                },
                "version_id": {
                    "type": "string",
                    "format": "uuid"
                }
            }
        },
...
        "request.TriggerConfig": {
            "type": "object",
            "properties": {
                "delay_ms": {
                    "type": "integer",
                    "example": 1500
                },
                "element_selector": {
                    "type": "string"
                },
                "inactivity_secs": {
                    "type": "integer",
                    "example": 30
                },
                "scroll_depth": {
                    "type": "integer",
                    "example": 50
                }
            }
        },
...
        "response.Audience": {
            "type": "object",
            "properties": {
                "max_shows": {
                    "type": "integer",
                    "example": 3
                },
                "only_new": {
                    "type": "boolean",
                    "example": false
                },
                "rules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.AudienceRule"
                    }
                },
                "show_once": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "response.GetCurrentAdminResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/response.CurrentAdmin"
                }
            }
        },
...
        "response.Hint": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "Здесь вы можете изменить параметры аккаунта"
                },
                "created_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "expected_input": {
                    "type": "string",
                    "example": "John Doe"
                },
                "id": {
                    "type": "string",
                    "format": "uuid",
                    "example": "01914f6a-8c9b-7a0b-9b0b-8c9b7a0b9b0c"
                },
                "input_placeholder": {
                    "type": "string",
                    "example": "Введите ваше имя"
                },
                "media_url": {
                    "type": "string",
                    "example": "https://cdn.example.com/onboarding/settings.mp4"
                },
                "page_path": {
                    "description": "Пусто — подсказка живёт на той же странице, где тур стартовал,\nто есть наследует target_path версии.",
                    "type": "string",
                    "example": "/dashboard"
                },
                "placement": {
                    "type": "string",
                    "enum": [
                        "top",
                        "bottom",
                        "center",
                        "left-top",
                        "right-top",
                        "left",
                        "right"
                    ],
                    "example": "bottom"
                },
                "required": {
                    "type": "boolean",
                    "example": false
                },
                "selector": {
                    "type": "string",
                    "example": "#settings-btn"
                },
                "spotlight": {
                    "type": "boolean",
                    "example": true
                },
                "step": {
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "example": "Нажмите кнопку настроек"
                },
                "tour_version_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "updated_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "wait_for_selector": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "response.Tour": {
            "type": "object",
            "properties": {
                "app_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "audience": {
                    "$ref": "#/definitions/response.Audience"
                },
                "created_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "description": {
                    "type": "string",
                    "example": "Тур по основным функциям дашборда"
                },
                "enabled": {
                    "type": "boolean",
                    "example": true
                },
                "hints": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.Hint"
                    }
                },
                "id": {
                    "type": "string",
                    "format": "uuid",
                    "example": "01914f6a-8c9b-7a0b-9b0b-8c9b7a0b9b0b"
                },
                "priority": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "draft",
                        "published",
                        "archived"
                    ],
                    "example": "draft"
                },
                "target_path": {
                    "type": "string",
                    "example": "/dashboard"
                },
                "title": {
                    "type": "string",
                    "example": "Добро пожаловать в систему"
                },
                "trigger_config": {
                    "$ref": "#/definitions/response.TriggerConfig"
                },
                "trigger_type": {
                    "type": "string",
                    "enum": [
                        "on_load",
                        "delay",
                        "exit_intent",
                        "manual",
                        "scroll_depth",
                        "inactivity",
                        "element_visible"
                    ],
                    "example": "on_load"
                },
                "updated_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "version": {
                    "type": "integer",
                    "example": 7
                },
                "version_id": {
                    "type": "string",
                    "format": "uuid"
                }
            }
        },
...
        "response.TriggerConfig": {
            "type": "object",
            "properties": {
                "delay_ms": {
                    "type": "integer",
                    "example": 1500
                },
                "element_selector": {
                    "type": "string"
                },
                "inactivity_secs": {
                    "type": "integer",
                    "example": 30
                },
                "scroll_depth": {
                    "type": "integer",
                    "example": 50
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "OIDC id token: \"Bearer \u003cid_token\u003e\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}